Adjusts imports for better consistency. Relocate some functions and modules for improved project organization.
