+++ hexabyte/actions/api/delete.py
@@ -3,11 +3,9 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from hexabyte.commands.command_parser import InvalidCommandError
-from hexabyte.constants.sizes import BYTE_BITS
-from hexabyte.utils.cursor import Cursor
-from hexabyte.utils.misc import str_to_int
-
+from ...commands import InvalidCommandError, str_to_int
+from ...constants.sizes import BYTE_BITS
+from ...cursor import Cursor
 from .._action import ActionError, UndoError
 from ._api_action import ReversibleApiAction
