+++ hexabyte/actions/api/open.py
@@ -4,8 +4,7 @@ from __future__ import annotations
 from pathlib import Path
 from typing import TYPE_CHECKING
 
-from hexabyte.commands.command_parser import InvalidCommandError
-
+from ...commands.command_parser import InvalidCommandError
 from .._action import ActionError
 from ._api_action import ApiAction
