+++ hexabyte/actions/api/replace.py
@@ -5,11 +5,9 @@ import struct
 from ast import literal_eval
 from typing import TYPE_CHECKING
 
-from hexabyte.commands.command_parser import InvalidCommandError
-from hexabyte.constants.sizes import BYTE_BITS
-from hexabyte.utils.context import context
-from hexabyte.utils.misc import int_fmt_str
-
+from ...commands import InvalidCommandError, int_fmt_str
+from ...constants.sizes import BYTE_BITS
+from ...context import context
 from .._action import ActionError, UndoError
 from ._api_action import ReversibleApiAction
