+++ hexabyte/actions/api/undo.py
@@ -3,9 +3,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from hexabyte.commands import InvalidCommandError
-from hexabyte.utils.misc import str_to_int
-
+from ...commands import InvalidCommandError, str_to_int
 from .._action import ActionError
 from ._api_action import ApiHandlerAction
