--- hexabyte/assets/__init__.py
@@ -1,1 +1,0 @@
-"""Hexabyte Asset Files."""
