--- hexabyte/assets/config.toml
@@ -1,57 +1,0 @@
-[general]
-max-cmd-history = 100
-max-undo        = 100
-plugins         = ["info", "entropy"]
-
-[normal]
-primary      = 'hex' # 'hex', 'bin', 'utf8'
-offset-style = 'hex' # 'hex', 'dec', 'off'
-
-# Column count is the number of columns per row
-# Column width represents the byte width of each column
-
-[normal.bin]
-column-count = 8
-column-size  = 1
-
-[normal.hex]
-column-count = 32
-column-size  = 1
-
-[normal.utf8]
-column-count = 1
-column-size  = 64
-
-[split]
-primary      = 'hex'  # 'hex', 'bin', 'utf8'
-secondary    = 'utf8' # 'hex', 'bin', 'utf8'
-offset-style = 'hex'  # 'hex', 'dec', 'off'
-
-[split.bin]
-column-count = 4
-column-size  = 1
-
-[split.hex]
-column-count = 4
-column-size  = 4
-
-[split.utf8]
-column-count = 8
-column-size  = 4
-
-[diff]
-primary      = 'hex' # 'hex', 'bin', 'utf8'
-secondary    = 'hex' # 'hex', 'bin', 'utf8'
-offset-style = 'hex' # 'hex', 'dec', 'off'
-
-[diff.bin]
-column-count = 4
-column-size  = 1
-
-[diff.hex]
-column-count = 4
-column-size  = 4
-
-[diff.utf8]
-column-count = 8
-column-size  = 4
