+++ hexabyte/view_components/__init__.py
@@ -1,5 +1,4 @@
 """Rich-compatible Components Package."""
-
 from .byte_view import ByteView
 from .hc_view import HCView
