+++ hexabyte/view_components/byte_view.py
@@ -13,10 +13,10 @@ from rich.style import Style
 from rich.text import Text
 from textual.geometry import Size
 
-from hexabyte.constants import DisplayMode
-from hexabyte.constants.sizes import BYTE_BITS, NIBBLE_BITS
-from hexabyte.utils.cursor import Cursor
-from hexabyte.utils.data_types import DataSegment
+from ..constants import DisplayMode
+from ..constants.sizes import BYTE_BITS, NIBBLE_BITS
+from ..cursor import Cursor
+from ..data_types import DataSegment
 
 NUMBERS_COLUMN_DEFAULT_PADDING = 3
